Use s:TRUE/s:FALSE instead of 1/0
vim-vimlparser can be translated into Python and JavaScript which
convert number to boolean implicitly, but it's difficlut to translate it
into other language which doesn't have implicit type conversion.

I actually created translater to Go (ref: #26) by including this change.
(We need more tweak for Go translater, but this change is one of them)

Using s:TRUE/s:FALSE is better not only for translator, but also for
readability.